always getting zero back from PDPID

+++ drwmutex/drwmutex_cpus.go
@@ -3,7 +3,7 @@
 
 package drwmutex
 
-func map_cpus() (cpus map[uint64]int) {
-	cpus = make(map[uint64]int)
+func map_cpus() (cpus map[int]int) {
+	cpus = make(map[int]int)
 	return
 }
